Feat: add product info
